fix(dev_react) refactored project structure
