--- src/utils/getAllItemsAPI.js
@@ -1,12 +1,0 @@
-export const getAllItems = async (data, size) => {
-    try {
-        const response = await fetch(`http://localhost:8080/item?size=${size}&page=${data-1}`);
-        if (!response.ok) {
-            throw new Error('Failed to fetch items');
-        }
-        return await response.json();
-    } catch (error) {
-        console.error(error);
-        throw error;
-    }
-}
